Finished core
Changed the detection column, so now it detects the row of pixels in the column closest to the robot. Which made it easier for the robot to handle turns and curves. Changed the speed at which the robot turns depending on whether it has a difference between the middle white line and the middle pixel. Added an extra else if statement that allows the robot to continue moving forward when it just about reaches the flag where it loses the white pixels.